requirements.txt


tensorflow
numpy
pillow
flask
gdown
